# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Juridica",
        "version": "1.0 (2024-07-22)"
    },
    "servers": [
        {
            "url": "http://localhost:8081/legal/public/"
        }
    ],
    "paths": {
...
        "/api/office": {
            "get": {
                "tags": [
                    "Office"
                ],
                "summary": "Get all Office",
                "operationId": "9c6bdf78328916c7a52a490ed702d975",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get account user information"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            },
            "post": {
                "tags": [
                    "Office"
                ],
                "summary": "Create Office",
                "operationId": "4f56e944a201491f6b9d37b65e0721c8",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "result": {
                                    "summary": "An result object.",
                                    "value": {
                                        "name": "",
                                        "address": ""
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/office/{id}": {
            "put": {
                "tags": [
                    "Office"
                ],
                "summary": "Update Office",
                "operationId": "f08798eaec05c1d994258540914c1f1e",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "type_doc",
                                    "number_doc"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Office 1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Office 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "Process"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Office"
                ],
                "summary": "Delete Office",
                "operationId": "b776666f2803ad6288d55be4ec0e3252",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            }
        },
...
    },
...
}